+++ docs/source/requirements.txt
@@ -1,1 +1,2 @@
 furo
+sphinx-argparse
