+++ docs/source/usage.rst
@@ -1,0 +1,11 @@
+Using `sby`
+===========
+
+Once SBY is installed and available on the command line as `sby`, either built from source or using
+one of the available CAD suites, it can be called as follows.  Note that this information is also
+available via `sby --help`.  For more information on installation, see :ref:`install-doc`.
+
+.. argparse::
+    :filename: ../sbysrc/sby_cmdline.py
+    :func: parser_func
+    :prog: sby
